requirements.txt
#python31012
Flask
pandas
secure-smtplib
py-emails
pycryptodomex
cryptography

